Working app, lots of changes
